Add ssh-agent to bash_profile, update linux setup

+++ linux-setup.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+paranoid=$([ $1 != "-y" ])
+
 dotfiles=(
   bash_profile
   bashrc
@@ -7,6 +9,7 @@ dotfiles=(
   gitconfig
   gitignore_global
   gvimrc
+  logout
   nethackrc
   screenrc
   vim
@@ -14,17 +17,35 @@ dotfiles=(
 )
 echo "Copying config files..."
 for f in ${dotfiles[@]}; do
-  if [[ -d $HOME/.$f ]]; then
-    read -p "$f already exists. Delete (y/N)? "
+  if [[ $paranoid && -d $HOME/.$f ]]; then
+    read -p "$f directory exists. Remove (y/N)? "
     if [[ "$REPLY" == "Y" || "$REPLY" == "y" ]]; then
       rm -rf $HOME/.$f
-      cp -Riv $f $HOME/.$f
+      cp -Rv $f $HOME/.$f
+    fi
+  elif [[ $paranoid && -e $HOME/.$f ]]; then
+    read -p "$f file exists. Overwrite (y/N)? "
+    if [[ "$REPLY" == "Y" || "$REPLY" == "y" ]]; then
+      cp -Rv $f $HOME/.$f
     fi
   else
-    cp -Riv $f $HOME/.$f
+      cp -Rv $f $HOME/.$f
   fi
 done
 
+echo
+echo "Configuring git..."
+git config --global core.excludesfile ~/.gitignore_global
+
+read -p "Enter your git user name: "
+if [ ! -z "$REPLY" ]; then
+  git config --global user.name "$REPLY"
+fi
+
+read -p "Enter your git user email: "
+if [ ! -z "$REPLY" ]; then
+  git config --global user.email "$REPLY"
+fi
 
 echo
 echo "Cloning the Vim plugins..."
@@ -49,15 +70,3 @@ for i in {0..4}; do
     git clone --quiet ${vimpluginurls[$i]} $HOME/.vim/bundle/${vimplugins[$i]}
   fi
 done
-
-
-echo
-echo "Configuring git..."
-git config --global core.excludesfile ~/.gitignore_global
-
-echo
-echo "Remember to set your git user name and email like so:"
-echo
-echo "    git config --global user.name \"Adam Dunson\""
-echo "    git config --global user.email \"ich@v0gel.us\""
-echo
